dev/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]==2.8.2
Pillow==10.4.0
slowapi==0.1.9
httpx==0.27.2

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Dev only
pytest==8.3.3
pytest-asyncio==0.24.0
